requirements.txt
bs4
requests
PyYAML == 5.3.1